+++ src/index.ts
@@ -7,4 +7,5 @@ export * from './parse';
 export * from './resolve';
 export * from './serialize';
 export * from './submit';
+export * from './visitor';
 export * from './validate';