+++ src/models/index.ts
@@ -4,4 +4,5 @@ export * from './embedded-link';
 export * from './entity';
 export * from './field';
 export * from './link';
+export * from './siren-element';
 export * from './sub-entity';