+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ### Added
 
+- Support for element traversal via the [Visitor pattern](https://en.wikipedia.org/wiki/Visitor_pattern) ([#14](https://github.com/siren-js/client/issues/14))
 - Support for validating fields on action submission ([#16](https://github.com/siren-js/client/issues/16))
 - Default JSON serializer ([#27](https://github.com/siren-js/client/issues/27))
 - Advanced `submit` usage examples the JSDocs